Move msgraph files to msgraph module instead of ms365. Improve provider documentation.
