+++ motor/providers/microsoft/msgraph/msgraphclient/adapter.go
