+++ motor/providers/microsoft/msgraph/msgraphconv/structs.go
